+++ library/runtime-ctrl-api/src/commonTest/kotlin/io/matthewnelson/kmp/tor/runtime/ctrl/api/address/LocalHostBaseTest.kt
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2024 Matthew Nelson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+package io.matthewnelson.kmp.tor.runtime.ctrl.api.address
+
+import kotlin.test.Test
+import kotlin.test.assertNotNull
+
+abstract class LocalHostBaseTest {
+
+    @Test
+    fun givenIPv4_whenResolved_thenIsCached() {
+        LocalHost.IPv4.resolve()
+        assertNotNull(LocalHost.IPv4.fromCache())
+    }
+
+    @Test
+    fun givenIPv6_whenResolved_thenIsCached() {
+        LocalHost.IPv6.resolve()
+        assertNotNull(LocalHost.IPv6.fromCache())
+    }
+}
